reorganized product section components
