+++ apps/main-landing/src/components/index.ts
@@ -3,3 +3,4 @@ export { Footer } from './footer';
 export { HeroSection } from './hero-section';
 export { SuperpowersSection } from './superpowers-section';
 export { TokenSection } from './token-section';
+export { ProductsSection } from './products-section'
