+++ apps/main-landing/src/components/products-section/components/index.ts
@@ -1,0 +1,2 @@
+export { DesktopProductsSection } from './desktop-products-section'
+export { MobileProductsSection } from './mobile-products-section'
