+++ apps/main-landing/src/components/products-section/components/mobile-products-section.tsx
@@ -1,8 +1,8 @@
 import { classes } from '@idriss-xyz/ui/utils';
 
-import { CreatorsSection } from './components/creators-section/creators-section';
-import { ExtensionSection } from './components/extension-section/extension-section';
-import { PredictionMarketsSection } from './components/prediction-markets-section/prediction-markets-section';
+import { CreatorsSection } from './creators-section';
+import { ExtensionSection } from './extension-section';
+import { PredictionMarketsSection } from './prediction-markets-section';
 
 type Properties = {
   className?: string;
